refactor(sqlite): Move sqlite extension traits to utils module
